mcp-toolbox/mcp-toolbox/orchestrator/requirements.txt
# Semantic Embedding Dependencies
gensim>=4.2.0
scikit-learn>=1.1.0
numpy>=1.21.0
scipy>=1.9.0

# Core Dependencies
chromadb>=0.4.0
requests>=2.28.0
python-dotenv>=0.19.0
PyMuPDF>=1.20.0

# Google Cloud (optional)
google-cloud-storage>=2.5.0
google-cloud-bigquery>=3.3.0
google-generativeai>=0.3.0

# Performance and utilities
asyncio-throttle>=1.0.0
aiofiles>=0.8.0

# AI/ML Libraries
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.12.0